fix: Fix regression where LocationBottomSheet wasn't shown
