+++ app/src/main/java/com/infomaniak/swisstransfer/ui/screen/newtransfer/upload/UploadFailureScreen.kt
@@ -19,6 +19,7 @@ package com.infomaniak.swisstransfer.ui.screen.newtransfer.upload
 
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.Image
+import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.State
@@ -27,12 +28,10 @@ import androidx.compose.ui.res.stringResource
 import com.infomaniak.core.compose.basics.CallableState
 import com.infomaniak.swisstransfer.BuildConfig
 import com.infomaniak.swisstransfer.R
-import com.infomaniak.swisstransfer.ui.components.BottomStickyButtonScaffold
-import com.infomaniak.swisstransfer.ui.components.BrandTopAppBar
-import com.infomaniak.swisstransfer.ui.components.EmptyState
-import com.infomaniak.swisstransfer.ui.components.LargeButton
+import com.infomaniak.swisstransfer.ui.components.*
 import com.infomaniak.swisstransfer.ui.images.AppImages.AppIllus
 import com.infomaniak.swisstransfer.ui.images.illus.appIntegrity.GhostScrollCrossPointing
+import com.infomaniak.swisstransfer.ui.images.illus.mascotSearching.MascotSearching
 import com.infomaniak.swisstransfer.ui.theme.SwissTransferTheme
 import com.infomaniak.swisstransfer.ui.utils.PreviewAllWindows
 import com.infomaniak.swisstransfer.upload.UploadState
@@ -52,6 +51,9 @@ fun UploadFailureScreen(
             exitNewTransfer = cancel,
             desc = stringResource(R.string.restrictedLocation)
         )
+    }
+    if (failure == UploadState.Failure.RestrictedLocation) {
+        LocationBottomSheet(closeButtonSheet = cancel)
     }
 }
 
@@ -78,6 +80,25 @@ private fun UploadFailureScreen(
             description = desc,
         )
     }
+}
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+private fun LocationBottomSheet(closeButtonSheet: () -> Unit) {
+    SwissTransferBottomSheet(
+        title = stringResource(R.string.sorry),
+        description = stringResource(R.string.restrictedLocation),
+        imageVector = AppIllus.MascotSearching.image(),
+        bottomButton = {
+            LargeButton(
+                title = stringResource(R.string.contentDescriptionButtonClose),
+                modifier = it,
+                style = ButtonType.Primary,
+                onClick = closeButtonSheet,
+            )
+        },
+        onDismissRequest = closeButtonSheet,
+    )
 }
 
 @PreviewAllWindows
